chore: move percentage chart to chart components
