+++ web/src/components/charts/PercentageBarChart/PercentageBar/PercentageBar.tsx
