+++ web/src/components/charts/PercentageBarChart/PercentageBar/index.tsx
