+++ web/src/components/charts/PercentageBarChart/PercentageBarChart.tsx
