+++ web/src/components/charts/PercentageBarChart/index.tsx
